+++ src/decode/mod.rs
@@ -9,6 +9,8 @@
 
 mod bc6h;
 mod bcn;
+#[cfg(all(feature = "simd", target_arch = "x86_64"))]
+mod simd;
 mod uncompressed;
 pub mod reference;